% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptIVPT.R
\docType{package}
\name{adaptIVPT}
\alias{adaptIVPT}
\title{adaptIVPT: Adaptive Bioequivalence Design for In-Vitro Permeation Tests for Pharmacokinetics with Mixed Scaling Approach}
\description{
This package helps design and analyze adaptive bioequivalence studies. Main functions are \code{msabe}, \code{rss}, \code{prms}, and \code{PRsurface}.
}
