% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.msabe.R
\name{summary.msabe}
\alias{summary.msabe}
\title{`summary` method for class "`msabe`"}
\usage{
\method{summary}{msabe}(object, ...)
}
\arguments{
\item{object}{an output from `msabe`}

\item{...}{additional arguments for print}
}
\value{
Does not return anything; print a summary of the output
}
\description{
`summary` method for class "`msabe`"
}
